Restructured into multiple subfolders.

+++ Maths/MathsUtil.c
@@ -2,7 +2,7 @@
 #include <math.h>
 
 float mathDegToRad(float deg) {
-	return deg / 360 * 2 * M_PI;
+	return (float)(deg / 360 * 2 * M_PI);
 }
 
 void mathCrossProduct(const float a[3], const float b[3], float cross[3]) {
@@ -12,7 +12,7 @@ void mathCrossProduct(const float a[3], const float b[3], float cross[3]) {
 }
 
 void mathNormVector(const float v[3], float norm[3]) {
-	float length = sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
+	float length = (float)sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
 	norm[0] = v[0] / length;
 	norm[1] = v[1] / length;
 	norm[2] = v[2] / length;
